requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
cachetools==7.0.6
certifi==2026.4.22
chardet==7.4.3
charset-normalizer==3.4.7
click==8.3.3
cssselect==1.4.0
cssutils==2.15.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
encutils==1.0.0
fastapi==0.136.1
frozenlist==1.8.0
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==0.1.20
langchain-classic==1.0.4
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.1.7
langchain-protocol==0.0.14
langchain-text-splitters==0.0.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.13
langsmith==0.1.147
lxml==6.1.0
marshmallow==3.26.2
more-itertools==11.0.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.109.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==23.2
premailer==3.10.0
propcache==0.4.1
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
xxhash==3.7.0
yagmail==0.15.293
yarl==1.23.0
zstandard==0.25.0
PyJWT==2.8.0
motor==3.7.1
